requirements.txt
# cbapi
requests 
colorama 
prettytable 
simplejson
PyCrypto
simple-crypt
pyaml
